academicworld_dash/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.3.0
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
neo4j==5.7.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
plotly==5.14.1
protobuf==3.20.3
pymongo==4.3.3
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
